+++ apps/api/src/schema/index.ts
@@ -4,7 +4,8 @@ export * from "./courses";
 export * from "./enrollment-history";
 export * from "./grades";
 export * from "./instructors";
+export * from "./larc";
 export * from "./search";
+export * from "./study-rooms";
 export * from "./websoc";
 export * from "./week";
-export * from "./study-rooms";
