+++ apps/api/src/schema/lib/index.ts
@@ -1,0 +1,3 @@
+export * from "./time.ts";
+export * from "./day.ts";
+export * from "./course.ts";
