+++ apps/api/src/services/larc.ts
@@ -1,10 +1,10 @@
+import type { larcQuerySchema, larcResponseSchema, larcSectionSchema } from "$schema";
 import type { database } from "@packages/db";
 import { type SQL, eq, getTableColumns, gte, ilike, lte, or } from "@packages/db/drizzle";
 import { and } from "@packages/db/drizzle";
 import { larcSection, websocCourse } from "@packages/db/schema";
 import { isTrue } from "@packages/db/utils";
 import type { z } from "zod";
-import type { larcQuerySchema, larcResponseSchema, larcSectionSchema } from "../schema/larc.ts";
 
 type LarcSessionServiceInput = z.infer<typeof larcQuerySchema>;
