+++ apps/data-pipeline/larc-scraper/src/lib.ts
@@ -5,7 +5,7 @@ import { larcSection, websocCourse } from "@packages/db/schema";
 import { parseStartAndEndTimes, sleep } from "@packages/stdlib";
 import { load } from "cheerio";
 import { fetch } from "cross-fetch";
-import { parseMeetingDays } from "../../../api/src/schema/lib/day-utils.ts";
+import { parseMeetingDays } from "../../../api/src/schema/lib/day.ts";
 
 type LarcSection = {
   daysString: string;
