+++ apps/data-pipeline/larc-scraper/worker-configuration.d.ts
@@ -1,6 +1,5 @@
 // Generated by Wrangler by running `wrangler types --x-include-runtime`
 
-export {};
-declare global {
-  const DB: Hyperdrive;
+interface Env {
+  DB: Hyperdrive;
 }
